Dirty hack. Disable home and screenshot buttons to bring total buttons down.
Found that some programs don't like the last d-pad button on B17 (0-index)
Not sure why, would make more sense if it had to fit in 16 bits, but wouldn't that
affect B16 as well? Anyway, the order seems to be determined by the event codes,
so the only way to bring the d-pad buttons down in the order is to remove some buttons.